Implementation of source-selection for audio as well as refinements for the export process -- avconv does not hang up and run slowly as the tasks have been seperated into two, [spec] probably improving pipelining.

+++ src/oli/bbp/Main.java
@@ -115,7 +115,6 @@ public class Main {
             SoundScheduler.instance = tfss;
             
             DimensionHelper.CURRENT_DOWNSCALE = 1.0;
-            
             while (OliRenderer.frameNum < OliRenderer.frameDur) {
                 String filename = directorySave.getPath() + File.separatorChar + String.format("%8d.png", OliRenderer.frameNum).replace(' ', '0');
                 OliRenderer.preprocessFrame();
@@ -123,47 +122,70 @@ public class Main {
                 Graphics2D g2d = bi.createGraphics();
                 OliRenderer.setupHints(g2d);
                 OliRenderer.renderFrame(g2d);
-                g2d.dispose();
-                ImageIO.write(bi, "png", new File(filename));
+                if (! OliRenderer.skipVid) {
+                    g2d.dispose();
+                    ImageIO.write(bi, "png", new File(filename));
+                }
                 if (OliRenderer.frameNum % REPORT_EVERY_FRAMES == 0) {
                     log.log(Level.INFO, "Frame: {0}", OliRenderer.frameNum);
                 }
             }
-            
             String outPathS = directorySave.getPath() + File.separatorChar;
             
             log.info("Finished frames, starting mixdown.");
             
             tfss.mixAndSave(outPathS + "audio.wav");
             
+            System.gc();
+            
             log.info("Finished mixdown.");
             
-            log.info("Will now attempt to run conversion with avconv");
-            
-            Runtime rt = Runtime.getRuntime();
-            String[] fcmd = new String[]{"avconv", "-r", Integer.toString(DimensionHelper.FRAMES_PER_SECOND), "-y", "-i", "./%08d.png", "-i", "./audio.wav", "-ar", "44100", "-strict", "experimental", "./result.mp4"};
-            String joinD = "";
-            for (String k : fcmd) {
-                joinD += " " + k;
-            }
-            log.log(Level.INFO, "Running:{0}", joinD);
-            Process p = rt.exec(fcmd, null, directorySave);
-            VoidStream vse = new VoidStream(p.getErrorStream());
-            VoidStream vso = new VoidStream(p.getInputStream());
-            vse.start();
-            vso.start(); // we don't want to know about stdout & stderr
-            int errc = p.waitFor();
-            if (errc != 0) {
-                throw new RuntimeException("avconv did not suceed, returned errcode " + errc);
-            } else {
-                log.log(Level.INFO, "Cleaning up...");
-                for (int i = 0; i < OliRenderer.frameDur; ++i) {
-                    String filename = directorySave.getPath() + File.separatorChar + String.format("%8d.png", i).replace(' ', '0');
-                    new File(filename).delete(); // could use deleteOnExit instead, but nah, might cause issues with many frames
-                }
-                new File(outPathS + "audio.wav").delete();
+            if (! OliRenderer.skipVid) {
+                log.info("Will now attempt to run conversion with avconv");
 
-                log.log(Level.INFO, "Cleaned up!");
+                Runtime rt = Runtime.getRuntime();
+                String[] fcmd = new String[]{"avconv", "-r", Integer.toString(DimensionHelper.FRAMES_PER_SECOND), "-y", "-i", "./%08d.png", "./video.ogv"};
+                String joinD = "";
+                for (String k : fcmd) {
+                    joinD += " " + k;
+                }
+                log.log(Level.INFO, "Running:{0}", joinD);
+                Process p = rt.exec(fcmd, null, directorySave);
+                VoidStream vse = new VoidStream(p.getErrorStream());
+                VoidStream vso = new VoidStream(p.getInputStream());
+                vse.start();
+                vso.start(); // we don't want to know about stdout & stderr
+                int errc = p.waitFor();
+                if (errc != 0) {
+                    throw new RuntimeException("avconv [0] did not suceed, returned errcode " + errc);
+                }
+                
+                rt = Runtime.getRuntime();
+                fcmd = new String[]{"avconv", "-y", "-i", "./video.ogv", "-i", "./audio.wav", "-strict", "experimental", "./result.mp4"};
+                joinD = "";
+                for (String k : fcmd) {
+                    joinD += " " + k;
+                }
+                log.log(Level.INFO, "Running:{0}", joinD);
+                p = rt.exec(fcmd, null, directorySave);
+                vse = new VoidStream(p.getErrorStream());
+                vso = new VoidStream(p.getInputStream());
+                vse.start();
+                vso.start(); // we don't want to know about stdout & stderr
+                errc = p.waitFor();
+                if (errc != 0) {
+                    throw new RuntimeException("avconv [1] did not suceed, returned errcode " + errc);
+                } else {
+                    log.log(Level.INFO, "Cleaning up...");
+                    for (int i = 0; i < OliRenderer.frameDur; ++i) {
+                        String filename = directorySave.getPath() + File.separatorChar + String.format("%8d.png", i).replace(' ', '0');
+                        new File(filename).delete(); // could use deleteOnExit instead, but nah, might cause issues with many frames
+                    }
+                    new File(outPathS + "audio.wav").delete();
+                    new File(outPathS + "video.ogv").delete();
+
+                    log.log(Level.INFO, "Cleaned up!");
+                }
             }
         } catch (InterruptedException ex) {
             Logger.getLogger(Main.class.getName()).log(Level.SEVERE, null, ex);
